requirements.txt
streamlit>=1.18.0
python-dotenv>=1.0.0
torch
torchvision
torchaudio
numpy
Pillow
opencv-python
matplotlib
scipy
scikit-learn
Altair
